% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper_functions.R
\name{make_topic_evo_concatenated}
\alias{make_topic_evo_concatenated}
\title{Function that adds a new RDS file named topic_evo_concatenated.RDS}
\usage{
make_topic_evo_concatenated(topic_evo_rds, directory = "inst/data/")
}
\arguments{
\item{topic_evo_rds}{the topic_evo.RDS as an object}

\item{directory}{the directory to save to}
}
\description{
Function that adds a new RDS file named topic_evo_concatenated.RDS
}
\examples{
\dontrun{
topic_evo_rds = readRDS("inst/data/topic_evo.RDS")
make_topic_evo_concatenated(topic_evo_rds)
}
}
